refactor(hir-ty): extract type compatibility
