+++ crates/hir-ty/src/infer.rs
@@ -5,11 +5,10 @@ use hir_def::{
     declaration::Declaration,
     def_id::DefId,
     expr::{
-        BinaryOp, Expr, ExprId, UnaryOp,
+        Expr, ExprId,
         data_ty::{BuiltinDataTy, DataTy, Dimension, IntKind, NamedDataTy},
         declarator::{DeclId, DeclaratorParent},
     },
-    literal::Literal,
     module::{ModuleId, ModuleKind, generate::GenerateBlockId, port::PortDeclId},
     pathres::{instance_target_def_id, resolve_name},
     stmt::{ForInit, StmtKind},
@@ -24,7 +23,7 @@ use utils::get::GetRef;
 use crate::{
     Type, TypeDiagnostic,
     db::TyDb,
-    ty::{BuiltinTy, Ty, TyClass, TyMember, TyResult},
+    ty::{BuiltinTy, Ty, TyMember, TyResult},
 };
 
 pub(crate) fn normalize_data_ty(
@@ -281,90 +280,6 @@ fn select_member(db: &dyn TyDb, base: &Ty, name: &Ident) -> TyResult {
         .find(|member| &member.name == name)
         .map(|member| TyResult::new(member.ty))
         .unwrap_or_else(|| TyResult::new(Ty::Unknown))
-}
-
-pub(crate) fn type_class(db: &dyn TyDb, ty: &Ty) -> Option<TyClass> {
-    match ty {
-        Ty::Alias { target, .. } => type_class(db, target),
-        Ty::Builtin(BuiltinTy::Data { id, .. }) => match db.lookup_intern_ty(*id) {
-            BuiltinDataTy::Int { .. } | BuiltinDataTy::Vector { .. } => Some(TyClass::Integral),
-            BuiltinDataTy::Real(_) => Some(TyClass::Real),
-            BuiltinDataTy::String => Some(TyClass::String),
-            BuiltinDataTy::Event | BuiltinDataTy::Chandle | BuiltinDataTy::Void => None,
-        },
-        Ty::Enum(_) => Some(TyClass::Integral),
-        Ty::Unknown
-        | Ty::Error
-        | Ty::Void
-        | Ty::Struct(_)
-        | Ty::Union(_)
-        | Ty::Queue { .. }
-        | Ty::Assoc { .. }
-        | Ty::Dynamic(_)
-        | Ty::Event
-        | Ty::Chandle
-        | Ty::Module(_)
-        | Ty::Checker(_)
-        | Ty::Covergroup(_)
-        | Ty::VirtualInterface { .. }
-        | Ty::GenerateBlock(_)
-        | Ty::Block(_) => None,
-    }
-}
-
-pub(crate) fn packed_bit_width(db: &dyn TyDb, ty: &Ty) -> Option<u64> {
-    match ty {
-        Ty::Alias { target, .. } => packed_bit_width(db, target),
-        Ty::Builtin(BuiltinTy::Data { id, container }) => match db.lookup_intern_ty(*id) {
-            BuiltinDataTy::String
-            | BuiltinDataTy::Real(_)
-            | BuiltinDataTy::Event
-            | BuiltinDataTy::Chandle
-            | BuiltinDataTy::Void => None,
-            BuiltinDataTy::Int { kind, .. } => Some(int_kind_width(kind) as u64),
-            BuiltinDataTy::Vector { dimensions, .. } => {
-                if dimensions.is_empty() {
-                    return Some(1);
-                }
-
-                let mut product: u64 = 1;
-                for dim in dimensions {
-                    let dim = dim?;
-                    let width = match dim {
-                        Dimension::Range(left, right) => {
-                            let left = eval_const_i128(db, *container, left)?;
-                            let right = eval_const_i128(db, *container, right)?;
-                            i128::abs(left - right).checked_add(1)?
-                        }
-                        Dimension::Size(size) => eval_const_i128(db, *container, size)?,
-                        Dimension::Queue(_) | Dimension::Assoc(_) | Dimension::Dynamic => {
-                            return None;
-                        }
-                    };
-                    let width: u64 = width.try_into().ok()?;
-                    product = product.checked_mul(width)?;
-                }
-                Some(product)
-            }
-        },
-        Ty::Unknown
-        | Ty::Error
-        | Ty::Void
-        | Ty::Struct(_)
-        | Ty::Enum(_)
-        | Ty::Union(_)
-        | Ty::Queue { .. }
-        | Ty::Assoc { .. }
-        | Ty::Dynamic(_)
-        | Ty::Event
-        | Ty::Chandle
-        | Ty::Module(_)
-        | Ty::Checker(_)
-        | Ty::Covergroup(_)
-        | Ty::VirtualInterface { .. }
-        | Ty::GenerateBlock(_)
-        | Ty::Block(_) => None,
-    }
 }
 
 fn normalize_data_ty_inner(
@@ -668,53 +583,6 @@ fn type_of_subroutine_port_impl(db: &dyn TyDb, port: InSubroutine<SubroutinePort
             )
         })
         .unwrap_or_else(|| TyResult::new(Ty::Unknown))
-}
-
-fn int_kind_width(kind: IntKind) -> usize {
-    match kind {
-        IntKind::Byte => 8,
-        IntKind::ShortInt => 16,
-        IntKind::Int => 32,
-        IntKind::LongInt => 64,
-        IntKind::Integer => 32,
-        IntKind::Time => 64,
-    }
-}
-
-fn eval_const_i128(db: &dyn TyDb, container: ArenaOwnerId, expr_id: ExprId) -> Option<i128> {
-    match expr_of(db, InContainer::new(container, expr_id))? {
-        Expr::Literal(Literal::Int(int)) => int.get_single_word().map(|v| v as i128),
-        Expr::Unary { op, expr } => {
-            let value = eval_const_i128(db, container, expr)?;
-            match op {
-                UnaryOp::Pos => Some(value),
-                UnaryOp::Neg => Some(value.checked_neg()?),
-                _ => None,
-            }
-        }
-        Expr::Binary { op, lhs, rhs } => {
-            let left = eval_const_i128(db, container, lhs)?;
-            let right = eval_const_i128(db, container, rhs)?;
-            match op {
-                BinaryOp::Add => left.checked_add(right),
-                BinaryOp::Sub => left.checked_sub(right),
-                BinaryOp::Mul => left.checked_mul(right),
-                BinaryOp::Div => (right != 0).then(|| left.checked_div(right)).flatten(),
-                BinaryOp::Mod => (right != 0).then(|| left.checked_rem(right)).flatten(),
-                BinaryOp::ShiftLeft => {
-                    u32::try_from(right).ok().and_then(|shift| left.checked_shl(shift))
-                }
-                BinaryOp::ShiftRight => {
-                    u32::try_from(right).ok().and_then(|shift| left.checked_shr(shift))
-                }
-                _ => None,
-            }
-        }
-        Expr::Cast { expr, .. } | Expr::SignedCast { expr, .. } => {
-            eval_const_i128(db, container, expr)
-        }
-        _ => None,
-    }
 }
 
 fn expr_of(db: &dyn TyDb, expr: InContainer<ExprId>) -> Option<Expr> {
