+++ crates/hir-ty/src/lib.rs
@@ -5,6 +5,7 @@
 //! new definition kind forces the type layer to classify it. This crate must
 //! not depend on semantic adapters or IDE features.
 
+mod compatibility;
 pub mod db;
 pub mod display;
 mod infer;
